test(testkit): move integration coverage and assert request state
