+++ crates/reinhardt-testkit/src/mock.rs
@@ -497,7 +497,6 @@ where
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use reinhardt_http::Handler;
 	use serde_json::json;
 
 	#[tokio::test]
@@ -591,25 +590,5 @@ mod tests {
 		spy.reset().await;
 		assert_eq!(mock.get_calls().await.len(), 0);
 		assert_eq!(spy.get_calls().await.len(), 0);
-	}
-
-	#[tokio::test]
-	async fn simple_handler_delegates_real_request_to_closure() {
-		// Arrange
-		let handler = SimpleHandler::new(|request: reinhardt_http::Request| {
-			Ok(reinhardt_http::Response::ok().with_body(request.uri.path().to_string()))
-		});
-		let request = reinhardt_http::Request::builder()
-			.uri("/mocked")
-			.body(bytes::Bytes::new())
-			.build()
-			.unwrap();
-
-		// Act
-		let response = handler.handle(request).await.unwrap();
-
-		// Assert
-		assert_eq!(response.status, http::StatusCode::OK);
-		assert_eq!(response.body, bytes::Bytes::from_static(b"/mocked"));
 	}
 }
